vulti-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vulti-agent"
version = "0.1.0"
description = "The self-improving AI agent — creates skills from experience, improves them during use, and runs anywhere"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Nous Research" }]
license = { text = "MIT" }
dependencies = [
  # Core
  "openai",
  "anthropic>=0.39.0",
  "python-dotenv",
  "fire",
  "httpx",
  "rich",
  "tenacity",
  "pyyaml",
  "requests",
  "jinja2",
  "pydantic>=2.0",
  # Interactive CLI (prompt_toolkit is used directly by cli.py)
  "prompt_toolkit",
  # Tools
  "firecrawl-py",
  "fal-client",
  # Text-to-speech (Edge TTS is free, no API key needed)
  "edge-tts",
  "faster-whisper>=1.0.0",
  # mini-swe-agent deps (terminal tool)
  "litellm>=1.75.5",
  "typer",
  "platformdirs",
  # Skills Hub (GitHub App JWT auth — optional, only needed for bot identity)
  "PyJWT[crypto]",
]

[project.optional-dependencies]
modal = ["swe-rex[modal]>=1.4.0"]
daytona = ["daytona>=0.148.0"]
dev = ["pytest", "pytest-asyncio", "pytest-xdist", "mcp>=1.2.0"]
messaging = ["python-telegram-bot>=20.0", "discord.py[voice]>=2.0", "aiohttp>=3.9.0", "slack-bolt>=1.18.0", "slack-sdk>=3.27.0"]
cron = ["croniter"]
slack = ["slack-bolt>=1.18.0", "slack-sdk>=3.27.0"]
cli = ["simple-term-menu"]
tts-premium = ["elevenlabs"]
voice = ["sounddevice>=0.4.6", "numpy>=1.24.0"]
pty = [
  "ptyprocess>=0.7.0; sys_platform != 'win32'",
  "pywinpty>=2.0.0; sys_platform == 'win32'",
]
honcho = ["honcho-ai>=2.0.1"]
mcp = ["mcp>=1.2.0"]
homeassistant = ["aiohttp>=3.9.0"]
acp = ["agent-client-protocol>=0.8.1,<1.0"]
rl = [
  "atroposlib @ git+https://github.com/NousResearch/atropos.git",
  "tinker @ git+https://github.com/thinking-machines-lab/tinker.git",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "wandb>=0.15.0",
]
yc-bench = ["yc-bench @ git+https://github.com/collinear-ai/yc-bench.git"]
all = [
  "vulti-agent[modal]",
  "vulti-agent[daytona]",
  "vulti-agent[messaging]",
  "vulti-agent[cron]",
  "vulti-agent[cli]",
  "vulti-agent[dev]",
  "vulti-agent[tts-premium]",
  "vulti-agent[slack]",
  "vulti-agent[pty]",
  "vulti-agent[honcho]",
  "vulti-agent[mcp]",
  "vulti-agent[homeassistant]",
  "vulti-agent[acp]",
  "vulti-agent[voice]",
]

[project.scripts]
vulti = "vulti_cli.main:main"
vulti-agent = "run_agent:main"
vulti-acp = "acp_adapter.entry:main"

[tool.setuptools]
py-modules = ["run_agent", "model_tools", "toolsets", "batch_runner", "trajectory_compressor", "toolset_distributions", "cli", "vulti_constants", "vulti_state", "vulti_time", "mini_swe_runner", "rl_cli", "utils"]

[tool.setuptools.packages.find]
include = ["agent", "tools", "tools.*", "vulti_cli", "gateway", "gateway.*", "cron", "honcho_integration", "acp_adapter"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring external services (API keys, Modal, etc.)",
]
addopts = "-m 'not integration' -n auto"
